started updating

+++ Sources/BeaconChain/DataStructures/Operations/Exit.swift
@@ -2,6 +2,6 @@ import Foundation
 
 struct Exit {
     let epoch: UInt64
-    let validatorIndex: UInt32 // @todo should be uint 24
+    let validatorIndex: UInt64
     let signature: Data
 }
